backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
python-multipart==0.0.20
Pillow==11.1.0
pytest==8.3.4
httpx==0.28.1
